requirements.txt
aiogram==2.16
aiohttp==3.8.1
aioscheduler==1.4.2
aiosignal==1.2.0
APScheduler==3.8.1
async-timeout==4.0.1
attrs==21.2.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.8
frozenlist==1.2.0
idna==3.3
multidict==5.2.0
psycopg2-binary==2.9.2
pyTelegramBotAPI==4.2.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
schedule==1.1.0
six==1.16.0
typing_extensions==4.0.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
yarl==1.7.2
